requirements.txt
alembic==1.0.3
asn1crypto==0.24.0
atomicwrites==1.2.1
attrs==18.2.0
bcrypt==3.1.4
blinker==1.4
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.4.1
coverage==4.5.2
coveralls==1.5.1
cryptography==2.4.2
docopt==0.6.2
dominate==2.3.4
Flask==1.0.2
Flask-API==1.0
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.7
Flask-Login==0.4.1
Flask-Mail==0.9.1
flask-marshmallow==0.9.0
Flask-Migrate==2.3.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
Flask-WTF==0.14.2
gunicorn==19.9.0
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
jwt==0.5.4
Mako==1.0.7
MarkupSafe==1.1.0
marshmallow==2.16.3
more-itertools==4.3.0
nose==1.3.7
pep8==1.7.1
Pillow==5.3.0
pluggy==0.8.0
psycopg2==2.7.6.1
py==1.7.0
pycparser==2.19
pytest==4.0.1
python-dateutil==2.7.5
python-editor==1.0.3
requests==2.20.1
six==1.11.0
SQLAlchemy==1.2.14
urllib3==1.24.1
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
